Cephadm: created separate test modules for bootstrap and host operations
